Update review component imports after folder reorganization
