+++ code/core/src/manager/components/review/components/ReviewCollectionPicker.tsx
@@ -2,12 +2,12 @@ import React, { useEffect, useRef, type FC } from 'react';
 
 import { styled } from 'storybook/theming';
 
-import { type StoryInfo } from './components/CollectionGrid.tsx';
 import {
   buildReviewStoryHref,
   prettifyComponentId,
   type ReviewNavEntry,
-} from './review-navigation.ts';
+} from '../review-navigation.ts';
+import { type StoryInfo } from './CollectionGrid.tsx';
 
 const PopoverList = styled.div(({ theme }) => ({
   display: 'flex',
