+++ code/core/src/manager/components/review/components/ReviewToolbarHeader.stories.tsx
@@ -2,21 +2,21 @@ import type { ReactNode } from 'react';
 
 import { expect, fn, within } from 'storybook/test';
 
+import { MemoryRouter } from 'storybook/internal/router';
 import {
   ManagerContext,
+  internal_fullStatusStore,
   type API,
   type State,
-  internal_fullStatusStore,
 } from 'storybook/manager-api';
-import { MemoryRouter } from 'storybook/internal/router';
 
-import preview from '../../../../../.storybook/preview.tsx';
-import { ADDON_ID, EVENTS } from './constants.ts';
+import preview from '../../../../../../.storybook/preview.tsx';
+import { ADDON_ID, EVENTS } from '../constants.ts';
+import { buildReviewChangesSummaryHref, buildReviewStoryHref } from '../review-navigation.ts';
+import type { ReviewState } from '../review-state.ts';
+import { useReviewShortcuts } from '../useReviewShortcuts.ts';
 import { ReviewProvider } from './ReviewProvider.tsx';
 import { ReviewToolbarHeader } from './ReviewToolbarHeader.tsx';
-import { buildReviewChangesSummaryHref, buildReviewStoryHref } from './review-navigation.ts';
-import type { ReviewState } from './review-state.ts';
-import { useReviewShortcuts } from './useReviewShortcuts.ts';
 
 type EventListener = (payload?: unknown) => void;
